test(heartbeat): slim to unique scheduler/config error-paths; cede RunnerCastHeartbeat coverage to dev's runnerCastHeartbeat.test.ts (swarm R1)
